Fixed: Problem with Firefox routing
When an external URL was being requested the new URL was still being stored in state, leading to the search app responding with a 'route not found' when it tried to load the state on a back button click.

This commit fixes this by only changing the state when the request is truly an internal one.

Fixes #3