Fixed: Help text was hidden in some places
Some help text was being hidden when it was rendered, due to overlapping elements.

This commit moves the application of the 'position' attribute to the facetHelp function, allowing the developer to set the position when creating the help text.

The position attribute is taken from elm-ui: https://package.elm-lang.org/packages/mdgriffith/elm-ui/latest/Element#nearby-elements