Refine search interface showing
This commit refines the process of showing the menu option. The previous option processed the entire button, but only chose to show it at the last stage of preparation.

This method optimizes the display to only actually render the choice when a national collection is not chosen. This removes the need for having a 'hidden' option on the button itself.