Fixed: nc choice resets to source view
Previously it was possible to select a person, choose a national collection, and then be left in limbo because people are not available to be searched in national collections.

This change fixes it so that when choosing a national collection the view is reset to showing the source search interface.