Fixed: ProbeData is now wrapped in a Response
Previously probe data was only a Maybe value. This commit changes the probe response to be a "Response data". This involves changing the Response type to accept a parameter of the type of data being wrapped.

This is the first step towards showing an animation when loading the probe results.

Refs rism-digital/rism-online-issues#130