Refactor records and search
Refactors the two main modules to more clearly separate them. They are now organized prinicpally around the return msg type. If there's a generic msg it usually means the view code is shared, so it can be moved to the UI module and shared between others. If there's a specific return msg (e.g., 'RecordMsg') then that code is tied to a specific module.

Part of an effort to reduce duplicated code.