Fixed: Rewrite routing handling
This introduces a rewrite of the routing handling for determining the search parameters. It fixes some issues with the national collection not being properly set in some cases, particularly when looking up numbers for the query results and the probe queries.

It also makes the process of initializing or updating various page models more consistent through the application of 'config' objects.

elm-review for unused imports and simplifications was also run.

Relates to rism-digital/rism-online-issues#169