Fixed: Don't show empty panels
If all controls are returning 'none' as a view, then we need to omit
the panel itself. This is done automatically for any controls that are
a direct descendent of the panel, but for cases where we want to reformat
the control layout directly, we need to 'manually' handle this.