Fixed: Loading bar gets stuck in some pages
This fixes an issue where the loading bar could get stuck since the
correct status of the loading process is not provided.

It fixes this by looking at whether we have a set of attached sources
and setting the status appropriately whether we'll need to show
the sources tab or not.