New: Static rendering
This implements the client side logic static rendering on first page load.

An initial data object can be passed into the Elm init flags that is used to load the data for that page. If this is null, then the page loads the data from an HTTP call (as it does now.)

Additionally the Country list data is passed in too.

A rework of the error response was also implemented so that we could combine JSON decoding errors (for the initial data) with HTTP decoding errors (from the HTTP loading.)